requirements.txt
aniso8601==8.0.0
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
Django==3.0rc1
django-crispy-forms==1.8.1
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wrapt==1.11.2